Properly handle sam header

+++ sinto/filterbarcodes.py
@@ -9,13 +9,26 @@ import os
 from itertools import chain
 
 
-def _iterate_reads(intervals, bam, cb, classes, trim_suffix, cellbarcode, readname_barcode):
+def _iterate_reads(
+    intervals,
+    bam,
+    cb,
+    classes,
+    trim_suffix,
+    cellbarcode,
+    readname_barcode,
+    sam,
+    outdir
+):
     inputBam = pysam.AlignmentFile(bam, "rb")
+    header = inputBam.header.to_dict()
+    newhead = dict((k, header[k]) for k in ("HD", "SQ", "RG"))
     ident = "".join(
         random.choice(string.ascii_uppercase + string.digits) for _ in range(6)
     )
-    filelist = [x + "_" + ident for x in classes]
-    bamlist = [pysam.AlignmentFile(x, "wb", template=inputBam) for x in filelist]
+    filemode = 'w' if sam else 'wb'
+    filelist = [os.path.join(outdir, x + "_" + ident) for x in classes]
+    bamlist = [pysam.AlignmentFile(x, filemode, header=newhead) for x in filelist]
     for i in intervals:
         for r in inputBam.fetch(i[0], i[1], i[2]):
             if readname_barcode is not None:
@@ -28,8 +41,8 @@ def _iterate_reads(intervals, bam, cb, classes, trim_suffix, cellbarcode, readna
                     cell_barcode = cell_barcode[:-2]
                 if cell_barcode in cb.keys():
                     cell_classes = cb[cell_barcode]
-                    for i in cell_classes:
-                        fileindex = filelist.index(i + "_" + ident)
+                    for j in cell_classes:
+                        fileindex = filelist.index(os.path.join(outdir, j + "_" + ident))
                         bamlist[fileindex].write(r)
     for i in bamlist:
         i.close()
@@ -37,22 +50,24 @@ def _iterate_reads(intervals, bam, cb, classes, trim_suffix, cellbarcode, readna
     return ident
 
 
-def mergeAll(idents, classes, nproc,  header, remove=True):
+def mergeAll(original, idents, classes, nproc, outdir, sam, remove=True):
     """Merge all temp files for each class
     If remove is True, remove temp files after
     successful merging"""
+    suffix = '.sam' if sam else '.bam'
     for i in classes:
-        allfiles = [i + "_" + x for x in idents]
-        output = i + '.bam'
-        pysam.merge('-@', str(nproc), '--no-PG', output, *allfiles)
-        if os.path.exists(output):
-            [os.remove(i) for i in allfiles]
-        else:
-            raise Exception("samtools merge failed, temp files not deleted")
+        allfiles = [os.path.join(outdir, i + "_" + x) for x in idents]
+        output = os.path.join(outdir, i + suffix)
+        pysam.merge('-@', str(nproc), '--no-PG', '-h', original, '-c', output, *allfiles)
+        if remove:
+            if os.path.exists(output):
+                [os.remove(i) for i in allfiles]
+            else:
+                raise Exception("samtools merge failed, temp files not deleted")
 
 
 def filterbarcodes(
-    cells, bam, readname_barcode, cellbarcode, sam=False, trim_suffix=True, nproc=1
+    cells, bam, readname_barcode, cellbarcode, outdir, sam=False, trim_suffix=True, nproc=1
 ):
     """Filter reads based on input list of cell barcodes
 
@@ -88,11 +103,7 @@ def filterbarcodes(
     unique_classes = list(set(chain.from_iterable(cb.values())))
     inputBam = pysam.AlignmentFile(bam, "rb")
     intervals = utils.chunk_bam(inputBam, nproc)
-    # write sam header
     ident = "".join(random.choice(string.ascii_uppercase + string.digits) for _ in range(6))
-    headerfile = "./" + ident + ".sam"
-    header = pysam.AlignmentFile(headerfile, "w", header = inputBam.header)
-    header.close()
     inputBam.close()
     if readname_barcode is not None:
         readname_barcode = re.compile(readname_barcode)
@@ -105,9 +116,18 @@ def filterbarcodes(
             classes=unique_classes,
             trim_suffix=trim_suffix,
             cellbarcode=cellbarcode,
-            readname_barcode=readname_barcode
+            readname_barcode=readname_barcode,
+            outdir=outdir,
+            sam=sam
         ),
         intervals.values(),
     ).get(9999999)
-    mergeAll(idents=idents, classes=unique_classes, nproc=nproc, header = headerfile, remove=True)
-    os.remove(headerfile)
+    mergeAll(
+        original=bam,
+        idents=idents,
+        classes=unique_classes,
+        nproc=nproc,
+        outdir=outdir,
+        sam=sam,
+        remove=True
+    )
